Primary confirmatory analysis

This function conducts the primary confirmatory analysis at the lastly passed checkpoint. In the analysis the mixed-effect logistic regression (see confirmatory_mixed_effect) and three Bayes factor analysis (see confirmatory_bayes_factor) are conducted. Final inference is made based on the inferences of the four conducted analyses.

analysis_confirmatory(df)

Arguments

df
dataframe, the input dataframe

Value

The output is a list containing the inputs, the results of the four analysis, and the inferences at the last checkpoint. The list contains:

highest_checkpoint, integer, the index of the investigated checkpoint

success, integer, the number of successful guesses

total_n, integer, the number of erotic trials

n_iteration, integer, the number of iteration (see confirmatory_mixed_effect)

mixed_effect_res, df, the results of the mixed-effect logistic regression

mixed_ci_width, numeric, width of the confidence interval

mixed_ci_lb, numeric, lower limit of the confidence interval 5

mixed_ci_ub, numeric, upper limit of the confidence interval 95

mixed_nhst_inference, character, inference based on the result

bf_res, df, the results of the three Bayes factor analysis

bf_replication, numeric, rounded Bf of analysis with prior based on Bem 2011 experiment 1 results

bf_uniform, numeric, rounded Bf of analysis with uniform prior

bf_buj, numeric, rounded Bf of analysis with BUJ prior

bf_inference, character, inference based on the three Bayes factors

inference, character, the summarized final inference

Examples

# \donttest{
# Running the confirmatory analysis
confirmatory_result <- analysis_confirmatory(df = example_m0)
# Checking the inferences at each checking point
confirmatory_result$inference

#> [1] "M0"

# }