Exploratory analysis

This function runs the exploratory analysis that aims to explore the empirical distribution of successful guess rate in the population, by contrasting it to a theoretical distribution expected if the true successful guess chance is homogeneously 50 potential irregularities in the distribution of guess rates.

analysis_exploratory(confirmatory_results)

Arguments

confirmatory_results
tibble, output of the analysis_confirmatory function

Value

The function returns a list of three, the possible success rates, the the proportion of observed successful guess rate, and the proportion of proportion of simulated successful guess rate.

Note

Exploratory analysis results will not affect the conclusions of our study.

Examples

# \donttest{
# Running the primary confirmatory analysis
confirmatory_results <- analysis_confirmatory(raw_data = example_m0)
# Running the exploratory analysis
analysis_exploratory(confirmatory_results = confirmatory_results)

#> $success_rates_theoretical_prop
#>           [,1]
#>  [1,] 0.000004
#>  [2,] 0.000088
#>  [3,] 0.000645
#>  [4,] 0.003098
#>  [5,] 0.011806
#>  [6,] 0.032570
#>  [7,] 0.070243
#>  [8,] 0.121475
#>  [9,] 0.166967
#> [10,] 0.185337
#> [11,] 0.167260
#> [12,] 0.121295
#> [13,] 0.071043
#> [14,] 0.032836
#> [15,] 0.011547
#> [16,] 0.003125
#> [17,] 0.000590
#> [18,] 0.000067
#> [19,] 0.000004
#> 
#> $success_rates_empirical_prop
#>               [,1]
#>  [1,] 0.0000000000
#>  [2,] 0.0000000000
#>  [3,] 0.0000000000
#>  [4,] 0.0039292731
#>  [5,] 0.0103143418
#>  [6,] 0.0422396857
#>  [7,] 0.0893909627
#>  [8,] 0.1208251473
#>  [9,] 0.1478388998
#> [10,] 0.1782907662
#> [11,] 0.1694499018
#> [12,] 0.1178781925
#> [13,] 0.0820235756
#> [14,] 0.0265225933
#> [15,] 0.0093320236
#> [16,] 0.0009823183
#> [17,] 0.0009823183
#> [18,] 0.0000000000
#> [19,] 0.0000000000
#> 
#> $possible_success_rates
#>  [1] 0.00 0.06 0.11 0.17 0.22 0.28 0.33 0.39 0.44 0.50 0.56 0.61 0.67 0.72 0.78
#> [16] 0.83 0.89 0.94 1.00
#> 

# }