Inference for the confirmatory analysis

This function calculates the final inference for the confirmatory analysis (see analysis_confirmatory). If all the four primary tests support M0 or M1 the final inference is conclusive. If after the last checkpoint not all the four tests support M0 or M1 the inference is inconclusive, otherwise the study is ongoing.

inference_confirmatory_combined(
  n_iteration,
  confirmatory_nhst_inference,
  confirmatory_bf_inference
)

Arguments

n_iteration
numeric, the number of iterations for the mixed effect analysis

confirmatory_nhst_inference
character, the inference based on the confirmatory NHST mixed analysis

confirmatory_bf_inference
character, the inference based on the Bayes factor analysis with 3 priors

Value

The function returns a character value that is either M1, M0, Inconclusive or Ongoing.

See also