Robustness test of Bf results with Bayesian parameter estimation

Robustness of BF results is tested by calculating HDI of the posterior distribution and checking its relation to the region of practical equivalence (ROPE), promoted in Kruschke, J. K., & Liddell, T. M. (2017). The Bayesian New Statistics: Hypothesis testing, estimation, meta-analysis, and power analysis from a Bayesian perspective. Psychonomic Bulletin & Review, 1-29.

robustness_bf_analysis(success, total_n)

Arguments

success

numeric, total number of successful guesses on erotic trials

total_n

numeric, total number of all erotic trials

Value

The function returns a list of five, the posterior density, the mode of the HDI, the upper and lower bound of the HDI and the inference that was based on these values (see inference_robustness_bf).